CONT::other: add fallback module for pidns-based detection
Add a special container detection module that runs after all other
modules and checks if the process is in a different PID namespace.
This module only implements check() and does not offer any restart
actions. It only serves to detect processes that are likely to belong
to containers, which are then omitted from regular output and are not
acted upon any further.